Relocated and compiled files
